chore(core): split our code from state_processor.go in state_processor_ext.go
